requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.121.2
frozenlist==1.8.0
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.7.0
openai==2.8.0
propcache==0.4.1
psycopg2-binary==2.9.9
pydantic==2.12.4
pydantic-settings==2.0.3
python-multipart==0.0.9
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.49.3
tqdm==4.67.1
twilio==9.8.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
